+++ AdventOfCode/Year2024/Day04/Solvers/Solver.cs
@@ -82,7 +82,7 @@ public class Solver : BaseSolver<string[]>
     {
         try
         {
-            return ParsedInput[point.Y][point.X];
+            return ParsedInput[(int)point.Y][(int)point.X];
         }
         catch { }
         return '.';
