+++ AdventOfCodeLibrary/Models/Vector3D.cs
@@ -1,11 +1,11 @@
 ﻿namespace AdventOfCodeLibrary.Models;
 public readonly struct Vector3D
 {
-    public int X { get; init; }
-    public int Y { get; init; }
-    public int Z { get; init; }
+    public long X { get; init; }
+    public long Y { get; init; }
+    public long Z { get; init; }
 
-    public Vector3D(int x, int y, int z)
+    public Vector3D(long x, long y, long z)
     {
         X = x;
         Y = y;
